debug

+++ detectron2/projects/Baseline/MA/network.py
@@ -1556,7 +1556,7 @@ class JointEstimationDisEmbedHead(DeepLabV3PlusHead):
         if self.loss_type == "panoptic_guided":
             get_gradient = Gradient(self.gradient_type)
             pan_targets = torch.unsqueeze(pan_targets, 1)
-
+            pan_targets = pan_targets.float()
             pan_targets_down = F.interpolate(pan_targets, scale_factor=0.25)
             pan_targets_down = pan_targets_down.float()
             pan_gradiant_x, pan_gradiant_y = get_gradient(pan_targets_down)
